# Remove waiting to check results of tracking tests.