--- a/spec/mocha.helpers.js
+++ b/spec/mocha.helpers.js
@@ -77,16 +77,6 @@ const extractBodyParamsFromFetch = (fetch) => {
   return null;
 };
 
-// Extract response parameters from event listener
-const extractResponseParamsFromListener = (listener) => {
-  const lastCallArguments = listener
-    && listener.args
-    && listener.args[listener.args.length - 1]
-    && listener.args[listener.args.length - 1][0];
-
-  return lastCallArguments;
-};
-
 module.exports = {
   setupDOM,
   teardownDOM,
@@ -95,5 +85,4 @@ module.exports = {
   clearStorage,
   extractUrlParamsFromFetch,
   extractBodyParamsFromFetch,
-  extractResponseParamsFromListener,
 };