--- a/spec/src/utils/request-queue.js
+++ b/spec/src/utils/request-queue.js
@@ -14,9 +14,10 @@ dotenv.config();
 
 describe('ConstructorIO - Utils - Request Queue', () => {
   const storageKey = '_constructorio_requests';
-  const waitInterval = 700;
+  const waitInterval = 1000;
   const requestQueueOptions = {
     sendTrackingEvents: true,
+    trackingSendDelay: 1,
   };
 
   describe('queue', () => {