Tracking tests are notoriously long to run and experience a lot of false negatives, as we're waiting for a set duration in order to check the status of the test.

This PR removes the wait, and simply checks the assertions once the "success" event es emitted. The result is:

- Less code
- Much, much faster test runs (about five times as fast)
- Fewer incidences of false positives

Before (8 failures):
```
npm run test  4.98s user 0.61s system 4% cpu 2:18.70 total
```

After (0 failures):
```
npm run test  4.28s user 0.61s system 17% cpu 28.375 total
```