@sblaurock 2021-06-24
> This looks great! The tests run way faster. There are 9 tests (all from utils) failing for me. Is this normal? Are we going to handle those separately?

@esezen good catch! I'll investigate those as well.